(notes) I think I identified a principal and actually fun way to identify non-standardized content with spacy matcher rule tokenization, should be reliably from bionomial proportion statistics (e.g. all 100 check out from a random sample, should be mega-good)
next steps should be to try this out.Since i'm really just tryign split on , conj things and skip certain kinds of uses of & (?)
this should be feasible. So, like a rule matcher that gets 80% and then a sstatistical matcher within that.
Note that you might no need full DEP parasing, there are lighter forms of generating DOCs