(WIP) Got really sidetracked on what I thought owuld be a simple regex extract on ' - ' where names includeing hyphens and dropping the last word of the name took me down a rabitt hole of negative/positive lookaheads and trying to capture up to a sequence of characters.
But, that said, I switched to trying to break on ' - ' and just taking the first element
but that's throwing some issues too. sheesh.

I got stuck here because if i got this done it would be the first field ready to deliver
standardized, whereas i'm supposed to investigate the acrive labeling server