(feat) Very crude active labeling notebook, "Process A," I suspect with the large number of labels that it might not do that great but this is what I wanted to get done for today and I was able to cobble it together well.
need to think through:
a) need to pass it context (the CIPCODE)
b) the real goal is to achieve saturation, which we do by having it show us labels that
it doesn't think it can predict. This might be the max/min active labeling strategy instead of what i have right now
which is between classes i think.
c) Normally, I would also want accuracy too but the reality is that once I have saturation I can
then turn to a more normal tool like Prod.igy and very quickly label it, easily export and stuff.