(verification) *(made a new git status, for ml?)*
Adds the program cousre name notebook, thought I added it, note the external data dump might spew into
the notebook folder, need to double check that

Ran through the first phase checking and was seeing ~90% accuracy for predicting if a name was good-enough or not.
THe jeffrye's prior was 0.88, which might just reflect it's bias away from 0 and 1.

That the jeffreys prior matches the active learn traiing so much makes me wonder if, well I suppose there has to be.
a relationship between the too. The learner was a ComplementNB whcih is supposed to handle minority classes well.

The next steps are to use the external datasets to replace the abbreviations in a giant search and replace fashion.
Oh there's a wikipedia page of abbreviations that might be even more useful too.