(feat/WIP) Implements .sql regex stream parsing, this was probablytoo much but requiring that a db and drivers be installed to parse a .sql file is also a big ask.
The other advantge here is that the parsing code skips processing of rows that do not have certain fields set
(like acroynmns), so the resulting data in memory is magnitudes smaller than if we stood up a database.

Need to test on actual .sql file, hopefully it'll process just like the two examples I tried on.